Try to fix threshold handling
Still a lot of work to do... Usually 0.96 is OK
So far I just added a warning if it is not reached
Is a 4% sacrifice of variance tolerable?

Should the standard error of the regression slope also be considered?
The error may not be very large since 95% of the variance
is explained by the fitting anyway